suppresses warnings under gcc 4.1.3.
git-svn-id: http://svn.berlios.de/svnroot/repos/sems/trunk@707 8eb893ce-cfd4-0310-b710-fb5ebe64c474

+++ core/plug-in/sipctrl/hash_table.cpp
@@ -476,7 +476,7 @@ void trans_bucket::dump()
     if(elmts.empty())
 	return;
 
-    DBG("*** Bucket ID: %i ***\n",get_id());
+    DBG("*** Bucket ID: %i ***\n",(int)get_id());
 
     for(trans_list::iterator it = elmts.begin(); it != elmts.end(); ++it) {
